requirements.txt
google-adk
google-cloud-firestore
google-genai
pydantic
python-dotenv